+++ packages/expo-application/ios/ApplicationModule.swift
@@ -11,7 +11,8 @@ public class ApplicationModule: Module {
         "applicationName": infoPlist?["CFBundleDisplayName"],
         "applicationId": infoPlist?["CFBundleIdentifier"],
         "nativeApplicationVersion": infoPlist?["CFBundleShortVersionString"],
-        "nativeBuildVersion": infoPlist?["CFBundleVersion"]
+        "nativeBuildVersion": infoPlist?["CFBundleVersion"],
+        "isLiquidGlassAvailable": ApplicationModule.isLiquidGlassAvailable()
       ]
     }
 
@@ -45,5 +46,20 @@ public class ApplicationModule: Module {
       let mainProvisioningProfile = EXProvisioningProfile.main()
       return mainProvisioningProfile.notificationServiceEnvironment()
     }
+  }
+
+  private static func isLiquidGlassAvailable() -> Bool {
+    #if compiler(>=6.2)  // Xcode 26
+    if #available(iOS 26.0, tvOS 26.0, macOS 26.0, *) {  // iOS 26
+      if let infoPlist = Bundle.main.infoDictionary {
+        // TODO(@uabx): Add a check for maximum SDK version when apple disables this flag
+        if let requiresCompatibility = infoPlist["UIDesignRequiresCompatibility"] as? Bool {
+          return !requiresCompatibility  // If the app requires compatibility then it will not use liquid glass
+        }
+      }
+      return true
+    }
+    #endif
+    return false
   }
 }